MAIN FILE READ DESCRIPTION
Listen, we'll make this the main FILE. first of all lets create all the
logic then we'll do the decorations. the new file uses fstream so now
use of that order array as well as we need to remove the structure and
make it into a class.
So let's do all our main work on it. k Bye